services/ai-service/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-service"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=0.1.0",
    "anthropic>=0.49.0",
]

[project.optional-dependencies]
dev = [
    "python-dotenv>=1.0.0",
]